backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2023.3

# Development
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1